@@ -1,9 +1,8 @@
 'use strict';
 
 var React = require('react');
-var ReactDOM = require('react-dom-polyfill')(React);
+var ReactDOM = require('react-dom');
 var findDOMNode = ReactDOM.findDOMNode;
-var className = require('classnames');
 var debounce = require('lodash.debounce');
 
 var CodeMirror = React.createClass({
@@ -11,7 +10,7 @@ var CodeMirror = React.createClass({
 
 	propTypes: {
 		className: React.PropTypes.any,
-		codeMirrorInstance: React.PropTypes.func,
+		codeMirrorInstance: React.PropTypes.string,
 		defaultValue: React.PropTypes.string,
 		onChange: React.PropTypes.func,
 		onFocusChange: React.PropTypes.func,
@@ -21,7 +20,12 @@ var CodeMirror = React.createClass({
 		value: React.PropTypes.string
 	},
 	getCodeMirrorInstance: function getCodeMirrorInstance() {
-		return this.props.codeMirrorInstance || require('codemirror');
+		return window[this.props.codeMirrorInstance];
+	},
+	getDefaultProps: function getDefaultProps() {
+		return {
+			codeMirrorInstance: 'CodeMirror'
+		};
 	},
 	getInitialState: function getInitialState() {
 		return {
@@ -82,10 +86,14 @@ var CodeMirror = React.createClass({
 		}
 	},
 	render: function render() {
-		var editorClassName = className('ReactCodeMirror', this.state.isFocused ? 'ReactCodeMirror--focused' : null, this.props.className);
+		var classes = 'ReactCodeMirror ';
+		if (this.state.isFocused) {
+			classes += 'ReactCodeMirror--focused';
+		}
+		classes += this.props.className;
 		return React.createElement(
 			'div',
-			{ className: editorClassName },
+			{ className: classes },
 			React.createElement('textarea', { ref: 'textarea', name: this.props.path, defaultValue: this.props.value, autoComplete: 'off' })
 		);
 	}
